Refactored and rewrote the fp classes.  The project now consists of three groups of classes (prefixed by "Free",  "FP", and "FPA"), and each one builds on the former:
(1)  The basic building block is the free module, maps between them and their elements.

(2)  A finitely presented (graded) module (in principle over any graded algebra) is represented by a morphism of finitely generated free modules from (1).  The class of fp. modules have functionality to compute kernels and images (plus related things).  The implementation of these functions are mostly concerned about performing linear algebra degree-wise in a range of dimensions.   If the algebra in question is finite, these functions terminate without help.  Otherwise, a parameter `top_dim` must be given to provide a termination condition.

(3)  The FPA_Module class and its associated homspace-, element- and morphism-class are specializations of the classes under (2) for when the algebra is one of the Steenrod algebras.  The code found in the FPA classes contain as little general module code as possible, and focuses on the idea that computations in fp modules over the Steenrod algebra can be done over a suitable finite sub-algebra.  Thus, most functions in this class follow the pattern:
  * Find a finite algebra A_fin for which the computtion can be done.
  * Change rings to A_fin.
  * Call a base class function in FP_Module.
  * Change rings for the result back to the original Steenrod algebra.

It is intended that the FPA-classes are to be made public for the user and that the rest of the classes should be considered private for the implementation.

(However, as a byproduct, the FP classes could be a good starting point if someone wants to develop a class for fp. graded modules over another ring than the Steenrod alegbra.)