Fix construction of zero element in free module and adopt directory naming convention.
The element_from_coordinates of the FreeModule class returns a sum
over a list. However, if this list is empty, sage returns an object
of the wrong type, namely an integer instead of a module element.
This is fixed by calling self._element_constructor_ on the result
of the sum.