The zero element should always have degree == None
Before, an FP_Element would have degree equal to the degree of the free element representing it.  This is correct if the element is non-zero, but wrong otherwise.  This bug is corrected so that degree() returns None if and only if the element is zero in the module.  As a consequence of the change made in the method degree(), the method _nonzero_() also had to be change in order to avoid having the two methods calling eachother.