Be strict about zero objects and their presentation.
This change cleans up how the method ``vector_presentation`` treat zero objects (both elements and homomorphisms).

The problem with presenting zero homomorphisms is that they do not have well defined degree.  Therefore, the codomain vectorspace cannot be determined and no presentation can be given.  For elements, the problem is that it is unknown what the containing vectorspace is.

Before, a zero homomorphism was presented as the trivial transformation by making an un-natural choice of degree for it.  This change avoids making it by refusing to present zero homomorphisms, as it should.

Before this change, zero elements were presented by the integer 0.  The coercion in Sage would make this work in practice since 0 is identified as the zero element of any vectorspace.  This change avoids this too, and returns ``None``, consistent with the corresponding method for presenting homomorphisms.