Condenced measurements
We are trying to separate between our code and external code, to be able to pin down what consumes most cycles when computing resolutions.
The code labelled "lin_alg" in the measurement output is external linear algebra.

Output of the test now looks like:

Computing f_1 (1/3)
Computing f_2 (2/3)
Resolving the kernel in the range of dimensions [1, 41]: 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40
Total time: 7.972053050994873s.

Time accounted for: 7.215564489364624s (90%)
                            lin_alg:    2.412s  33% of accounted time.
        self.vector_presentation(n):    1.379s  19% of accounted time.
           element_from_coordinates:     0.24s  3% of accounted time.
           j.vector_presentation(n):    3.185s  44% of accounted time.

 41.
Computing f_3 (3/3)
Resolving the kernel in the range of dimensions [2, 41]: 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40
Total time: 9.619606971740723s.

Time accounted for: 9.604811906814575s (99%)
                            lin_alg:    2.835s  29% of accounted time.
        self.vector_presentation(n):    3.233s  33% of accounted time.
           element_from_coordinates:    0.348s  3% of accounted time.
           j.vector_presentation(n):    3.189s  33% of accounted time.

 41.